Add error responses for bad params to GET /nodes
GET requests to the /nodes endpoint will now return a HTTP 400 - Invalid
Request response.

This also fixes a bug caused by improper use of the operator "in".
I assumed that in worked on Sets as you would expect in python. However,
I believe it works on Sets in a way that they are objects. So in isn't
a compile time bug when used on a Set, but it is a logic bug, because
the Set object will be checked for the keyword, not the actual
container.