/nodes: Add POST support to add node(s)
This commit adds support for sending POST commands to the /nodes
endpoint. POSTs with invalid formatting of data will receive
a BAD_REQUEST response along with a stringified message of the exception
which caused the error. This could be a security hole at some point and
should probably be addressed, but for now we hope no unexpected
exceptions will arise with sensitive data.

This commit has reduced code quality by placing all class definitions in
the nodes.controller.ts file. This was done to improve development speed
but should be undone in a future commit.