Add support for GET /nodes/:id/data
GET requests to /nodes/:id/data  now return a list of keys to the data
component of a node.

This change also caused us to make an internal change. "data" is now
automatically added as an attribute of nodes internally.

This also fixes a bug in the returned response for GET /nodes. The
response no longer contains a dictionary with
{ "metadata": <result-dict> }, it now contains just the expected result
dictionary.