Add /scheduler/(kill)|(wait) endpoints
This commit adds support for waiting for a set state by using the GET
/scheduler/wait endpoint. It also adds support for setting the state to
"killing" via a PUT to the /scheduler/kill endpoint.

Work on this commmit is in a very poor state. All the work for the
/scheduler endpoint needs to be encapsulated into its own controller
class.

Note, this also adds a dependency on node version >=11.13.0 since we use
the events.once async function provided.